refactor: moved k8s dev mode config to k8s config module
