+++ core/src/plugins/kubernetes/config.ts
@@ -6,41 +6,99 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import dedent = require("dedent")
-
 import {
+  joi,
   joiArray,
   joiIdentifier,
+  joiIdentifierDescription,
   joiProviderName,
-  joi,
+  joiSparseArray,
   joiStringMap,
   StringMap,
-  joiIdentifierDescription,
-  joiSparseArray,
 } from "../../config/common"
-import { Provider, providerConfigBaseSchema, BaseProviderConfig } from "../../config/provider"
+import { BaseProviderConfig, Provider, providerConfigBaseSchema } from "../../config/provider"
 import {
-  containerRegistryConfigSchema,
-  ContainerRegistryConfig,
-  commandExample,
-  containerEnvVarsSchema,
-  containerArtifactSchema,
-  ContainerEnvVars,
   artifactsDescription,
+  commandExample,
+  containerArtifactSchema,
+  containerDevModeSchema,
+  ContainerDevModeSpec,
+  ContainerEnvVars,
+  containerEnvVarsSchema,
+  ContainerRegistryConfig,
+  containerRegistryConfigSchema,
+  syncDefaultDirectoryModeSchema,
+  syncDefaultFileModeSchema,
+  syncDefaultGroupSchema,
+  syncDefaultOwnerSchema,
+  syncExcludeSchema,
 } from "../container/config"
 import { PluginContext } from "../../plugin-context"
-import { deline } from "../../util/string"
+import { dedent, deline } from "../../util/string"
 import { defaultSystemNamespace } from "./system"
-import { hotReloadableKinds, HotReloadableKind } from "./hot-reload/hot-reload"
-import { baseTaskSpecSchema, BaseTaskSpec, cacheResultSchema } from "../../config/task"
-import { baseTestSpecSchema, BaseTestSpec } from "../../config/test"
+import { HotReloadableKind, hotReloadableKinds } from "./hot-reload/hot-reload"
+import { BaseTaskSpec, baseTaskSpecSchema, cacheResultSchema } from "../../config/task"
+import { BaseTestSpec, baseTestSpecSchema } from "../../config/test"
 import { ArtifactSpec } from "../../config/validation"
 import { V1Toleration } from "@kubernetes/client-node"
 import { runPodSpecIncludeFields } from "./run"
-import { KubernetesDevModeDefaults, kubernetesDevModeDefaultsSchema } from "./dev-mode"
 import { KUBECTL_DEFAULT_TIMEOUT } from "./kubectl"
+import { devModeGuideLink } from "./dev-mode"
 
 export const DEFAULT_KANIKO_IMAGE = "gcr.io/kaniko-project/executor:v1.6.0-debug"
+
+export interface KubernetesDevModeSpec extends ContainerDevModeSpec {
+  containerName?: string
+}
+
+export interface KubernetesDevModeDefaults {
+  exclude?: string[]
+  fileMode?: number
+  directoryMode?: number
+  owner?: number | string
+  group?: number | string
+}
+
+export const kubernetesDevModeSchema = () =>
+  containerDevModeSchema().keys({
+    containerName: joiIdentifier().description(
+      `Optionally specify the name of a specific container to sync to. If not specified, the first container in the workload is used.`
+    ),
+  }).description(dedent`
+    Specifies which files or directories to sync to which paths inside the running containers of the service when it's in dev mode, and overrides for the container command and/or arguments.
+
+    Note that \`serviceResource\` must also be specified to enable dev mode.
+
+    Dev mode is enabled when running the \`garden dev\` command, and by setting the \`--dev\` flag on the \`garden deploy\` command.
+
+    See the [Code Synchronization guide](${devModeGuideLink}) for more information.
+  `)
+/**
+ * Provider-level dev mode settings for the local and remote k8s providers.
+ */
+export const kubernetesDevModeDefaultsSchema = () =>
+  joi.object().keys({
+    exclude: syncExcludeSchema().description(dedent`
+        Specify a list of POSIX-style paths or glob patterns that should be excluded from the sync.
+
+        Any exclusion patterns defined in individual dev mode sync specs will be applied in addition to these patterns.
+
+        \`.git\` directories and \`.garden\` directories are always ignored.
+      `),
+    fileMode: syncDefaultFileModeSchema(),
+    directoryMode: syncDefaultDirectoryModeSchema(),
+    owner: syncDefaultOwnerSchema(),
+    group: syncDefaultGroupSchema(),
+  }).description(dedent`
+    Specifies default settings for dev mode syncs (e.g. for \`container\`, \`kubernetes\` and \`helm\` services).
+
+    These are overridden/extended by the settings of any individual dev mode sync specs for a given module or service.
+
+    Dev mode is enabled when running the \`garden dev\` command, and by setting the \`--dev\` flag on the \`garden deploy\` command.
+
+    See the [Code Synchronization guide](${devModeGuideLink}) for more information.
+  `)
+
 export interface ProviderSecretRef {
   name: string
   namespace: string
